Fix Integer to Long type conversion for quantity fields
- DailySummary: totalOrders Integer -> Long
- Inventory: quantity Integer -> Long, version add columnDefinition
- OrderDetail: quantity Integer -> Long
- OrderService.OrderDetailRequest: quantity Integer -> Long
- OrderMapper: method parameters Integer -> Long
- OrderWebController.OrderForm: quantities List<Integer> -> List<Long>
- OracleFeatureTestController: test method parameters Integer -> Long

Reason: Oracle NUMBER type converted to PostgreSQL BIGINT by DMS

+++ src/main/java/com/autoever/mes/domain/history/entity/DailySummary.java
@@ -21,7 +21,7 @@ public class DailySummary {
     private LocalDateTime summaryDate;
     
     @Column(name = "TOTAL_ORDERS")
-    private Integer totalOrders;
+    private Long totalOrders;
     
     @Column(name = "TOTAL_AMOUNT", precision = 15, scale = 2)
     private BigDecimal totalAmount;
