+++ src/main/java/com/autoever/mes/domain/order/service/OrderService.java
@@ -86,11 +86,11 @@ public class OrderService {
     
     public static class OrderDetailRequest {
         private Long productId;
-        private Integer quantity;
+        private Long quantity;
         
         public Long getProductId() { return productId; }
         public void setProductId(Long productId) { this.productId = productId; }
-        public Integer getQuantity() { return quantity; }
-        public void setQuantity(Integer quantity) { this.quantity = quantity; }
+        public Long getQuantity() { return quantity; }
+        public void setQuantity(Long quantity) { this.quantity = quantity; }
     }
 }
